fix(user-store): rename user => users [WiP]
*	modified:   apps/nx-angular-next/src/app/core/core.module.ts
*	modified:   libs/frontend/shared/data-access/user-store/src/index.ts
*	renamed:    libs/frontend/shared/data-access/user-store/src/lib/+state/user.effects.ts -> libs/frontend/shared/data-access/user-store/src/lib/+state/users.effects.ts
*	renamed:    libs/frontend/shared/data-access/user-store/src/lib/+state/user.reducer.ts -> libs/frontend/shared/data-access/user-store/src/lib/+state/users.reducer.ts
*	renamed:    libs/frontend/shared/data-access/user-store/src/lib/+state/user.selectors.ts -> libs/frontend/shared/data-access/user-store/src/lib/+state/users.selectors.ts
*	modified:   libs/frontend/shared/data-access/user-store/src/lib/interfaces/users-apollo.interface.ts
*	modified:   libs/frontend/shared/data-access/user-store/src/lib/interfaces/users-store-feature-key.interface.ts
*	renamed:    libs/frontend/shared/data-access/user-store/src/lib/services/user-apollo.service.ts -> libs/frontend/shared/data-access/user-store/src/lib/services/base-user-apollo.service.ts
*	renamed:    libs/frontend/shared/data-access/user-store/src/lib/+state/user.facade.ts -> libs/frontend/shared/data-access/user-store/src/lib/services/base-users-facade.service.ts
*	renamed:    libs/frontend/shared/data-access/user-store/src/lib/user-store.module.ts -> libs/frontend/shared/data-access/user-store/src/lib/users-store.module.ts
*	modified:   libs/shared/utils/store/src/lib/effects/abstract-effects.ts
