+++ libs/frontend/shared/data-access/user-store/src/index.ts
@@ -1,6 +1,6 @@
-export * from './lib/+state/user.actions';
-export * from './lib/+state/user.reducer';
-export * from './lib/+state/user.selectors';
-export * from './lib/+state/user.models';
-export * from './lib/+state/user.facade';
-export * from './lib/user-store.module';
+export * from './lib/+state/users.actions';
+export * from './lib/+state/users.reducer';
+export * from './lib/+state/users.selectors';
+export * from './lib/services/base-users-facade.service';
+export * from './lib/users-store.module';
+export * from './lib/interfaces/users-facade.interface';
