+++ libs/frontend/shared/data-access/user-store/src/lib/interfaces/users-apollo.interface.ts
@@ -1,4 +1,4 @@
-import { IUser, TApolloResponse } from '@svvs/shared/utils/interfaces';
+import { TApolloResponse } from '@svvs/shared/utils/interfaces';
 
 /**
  * User Apollo
@@ -7,7 +7,5 @@ export abstract class IUsersApollo {
   /**
    * Load current user
    */
-  abstract loadUser(
-    queryParams?: Record<string, unknown>
-  ): TApolloResponse<IUser>;
+  abstract loadUser(queryParams?: Record<string, unknown>): TApolloResponse;
 }
