+++ libs/frontend/shared/data-access/user-store/src/lib/interfaces/users-store-feature-key.interface.ts
@@ -1,4 +1,4 @@
-import { USERS_FEATURE_KEY } from '../+state/user.reducer';
+import { USERS_FEATURE_KEY } from '../+state/users.reducer';
 import { IUsersState } from './users-state.interface';
 
 export interface IUsersStoreFeatureKey {
